backend/requirements.txt
django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
uvicorn==0.29.0
psycopg[binary]==3.1.18
pillow==10.3.0
